+++ contracts/implementations/HashCrashERC20.sol
@@ -13,8 +13,8 @@ contract HashCrashERC20 is HashCrash {
         address lootTable_,
         bytes32 genesisHash_,
         address hashProducer_,
-        uint128 maxExposureNumerator_,
-        uint128 lowLiquidityThreshold_,
+        uint64 maxExposureNumerator_,
+        uint256 lowLiquidityThreshold_,
         address owner_,
         address token_,
         uint256 minimumValue_
