+++ contracts/test/HashCrashHarness.sol
@@ -9,8 +9,8 @@ contract HashCrashHarness is HashCrash {
         address lootTable_,
         bytes32 genesisHash_,
         address hashProducer_,
-        uint128 maxExposureNumerator_,
-        uint128 lowLiquidityThreshold_,
+        uint64 maxExposureNumerator_,
+        uint256 lowLiquidityThreshold_,
         address owner_,
         address token_,
         uint256 minimumValue_
